+++ test/test_helper.rb
@@ -2,7 +2,7 @@ $LOAD_PATH.unshift File.expand_path("../lib", __dir__)
 require "trailblazer/developer"
 
 require "minitest/autorun"
-
+require "pp"
 
 require "trailblazer/activity"
 require "trailblazer/activity/testing"