+++ lib/trailblazer/developer/CHANGES.md
@@ -1,3 +1,17 @@
+# 0.0.6
+
+* Remove ID extraction logic from `Generate`, this is done on the server-side.
+
+# 0.0.5
+
+* Introduce `:query` option for `Client.import`.
+* In `Generate`, added `Element.parent` field.
+* No more magic is applied when extracting the ID/semantic. We just use the plain string.
+
+# 0.0.4
+
+* Allow injecting `:parser` into `Generate.transform_from_hash`.
+
 # 0.0.3
 
 * Add `Developer.railway`.